==> notebooks/01 Data Scraping.ipynb <==
# ✅Step 1: Data Scraping

## 0. 🎯Import libraries

import sys
import requests as r

import pandas as pd
from datetime import datetime

from tqdm import tqdm

# Import our own modules
sys.path.append("../scripts/")
import chadtools

## 1. 🎯Authenticate with Reddit API

Using a function defined in our `utils.py` script, we can authenticate with the Reddit API using our own `credentials.json` file, and get a `dict` of headers to be used in all subsequent GET requests.

headers = chadtools.authenticate_and_get_headers()
headers

## 2. 🎯Sending our GET requests

### 2.1 Prepare GET request for all Flairs + Paginate through all search results

To get all the posts from the subreddit, we will iterate through a list of all the flairs in the subreddit and send a GET request for each flair. We will then paginate through the search results using the `after` ID given by Reddit's API to get all the posts for each flair.

s = r.Session()
BASE_ENDPOINT = "https://oauth.reddit.com"
flair_names = ['Recipe', 'Dessert', 'Pasta', 'Poultry', 'Vegetarian', 'Drink', 'Beef', 'Pork', 'Seafood', 'Fruit\Vegetarian']
subreddit_name = 'recipes'

all_data_for_all_flairs = []
all_data_by_flair = {}

for flair in tqdm(flair_names):
    flair_query = f'flair_name:"{flair}"'
    params = {
        'q': flair_query,
        'limit': 100,
        'restrict_sr': 1,
        'sort': 'new'
    }
    response = s.get(f"{BASE_ENDPOINT}/r/{subreddit_name}/search", params=params, headers=headers)
    # Initialize an empty list to store the data from page for the current flair
    all_data_by_flair[flair] = []
    
    # Process the data from the first page
    data = response.json()
    all_data_by_flair[flair].extend(data['data']['children'])

    # Page 02 and beyond
    while 'after' in data['data'] and data['data']['after'] is not None:
        after_id = data['data']['after']
        params["after"] = after_id
        response = s.get(f"{BASE_ENDPOINT}/r/{subreddit_name}/search", params=params, headers=headers)
        # print(f"Requesting Page {len(all_data_by_flair[flair]) // params['limit'] + 1}")
        data = response.json()

        # Process the data from the current page
        #all_data_by_flair.extend(data['data']['children'])
        all_data_by_flair[flair].extend(data['data']['children'])
    
    all_data_for_all_flairs.extend(all_data_by_flair[flair])
    
len(all_data_for_all_flairs)

## 3. 🎯Saving the data 

### 3.1 Create a dataframe of all posts 

df_posts = pd.DataFrame(all_data_for_all_flairs)
df_posts = pd.json_normalize(df_posts['data'], max_level=0)

df_posts['permalink'] = "https://reddit.com" + df_posts['permalink']
df_posts.tail()

df_posts.info()

### 3.2 Save dataframe as a JSON file

df_posts.to_json('../data/posts.json', orient='records', indent=4)

### 3.3 Saving Dataframe as HTML for webpage

df_posts = pd.read_json('../data/posts.json')

df_posts_styled = df_posts.head().style.set_table_styles([
    {'selector': 'thead', 'props': [('background-color', '#f2f2f2')]},
    {'selector': 'th', 'props': [('border', '1px solid #dddddd')]},
    {'selector': 'td', 'props': [('border', '1px solid #dddddd')]},
])

df_posts_styled.to_html('../docs/posts.html', render_links=True, index=False)
