==> notebooks/03 Data cleaning.ipynb <==
# ✅Step 3: Data Cleaning

This notebook will clean up the posts.json file by:
1. Removing posts with non-English titles
2. Removing irrelevant columns
3. Converting relevant values to the correct data types

We will then use GPT-3.5 to extract a list of food ingredients from the recipe comment.

## 0. 🎯Setup

import sys

import pandas as pd
from datetime import datetime

from tqdm import tqdm
tqdm.pandas()

# Import our own modules
sys.path.append("../scripts/")
import chadtools

### 0.1. Load json file 

Load dataframe from json file generated from [02 Scraping OP Comments](./02%20Scraping%20OP%20comment.ipynb)


df = pd.read_json('../data/posts_with_comments.json', orient='records')
df.shape

## 1. 🧹Simple Data Cleanup

Some posts are not formatted properly or have been deleted. We will remove them from our dataframe by checking for newlines, which are present in all properly formatted recipes.

df = df[df['recipe_comment'].str.contains("\n") == True]

df.shape

We clean up the dataframe by filtering for the columns that we want.

desired_columns = ['id',
                   'title',
                   'score',
                   'num_comments',
                   'recipe_comment',
                   'created_utc', 
                   'upvote_ratio',
                   'link_flair_text',
                   'author',
                   'url',
                   'comment_link',
                   'permalink']

df_filtered = df.loc[:, desired_columns]
df_filtered.head()

df_filtered.shape

We filter out posts with empty values for the `id`, `title`, `recipe_comment` and `permalink` columns.

important_cols = ['id', 'title', 'recipe_comment', 'permalink']
df_filtered.dropna(axis=0, subset=important_cols, inplace=True)
df_filtered.shape

### 1.1 Filter out posts with non-english titles

# filter the english posts by applying custom function
df_filtered = df_filtered[df_filtered['title'].progress_apply(chadtools.is_english, threshold_rank=5)]

df_filtered.shape

### 1.2 Convert data to more appropriate types 

df_filtered.info()

We will convert the following columns to the following types:
- `title` to string
- `score` to int32
- `num_comments` to int16
- `created_utc` to datetime
- `upvote_ratio` to float16

#### 1.2.1 Convert created_utc to a datetime object

df_filtered['created_utc'] = df_filtered['created_utc'].apply(lambda x: datetime.fromtimestamp(x))

Starting from 31 Aug 2020, r/recipes imposed strict rules on the format of new posts, which made the format substantially more consistent. Hence, we will only keep posts from 31 Aug 2020 onwards for ease of data collection.

cutoff_datetime = pd.to_datetime("2020-08-31 00:00:00")

# Filter out rows where 'created_utc' is before the cutoff datetime
df_filtered = df_filtered[df_filtered['created_utc'] >= cutoff_datetime]

df_filtered.tail()

#### 1.2.2 Convert the more primitive types to more size-efficient types.

df_filtered['score'] = df_filtered['score'].astype('int16')
df_filtered['num_comments'] = df_filtered['num_comments'].astype('int16')   
df_filtered['upvote_ratio'] = df_filtered['upvote_ratio'].astype('float16')

df_filtered.info()

As seen, the size of the dataframe has been reduced by **50%**.

### Save df_filtered as a JSON file

df_filtered.to_json('../data/cleaned_posts.json', indent=4, orient='records')

### Saving as HTML to add to Index.md

df_filtered.head().to_html('../docs/df_filtered.html')

## Identify Cuisines and Ingredients

We will use GPT-3.5 to identify the cuisine of each recipe and extract a list of ingredients.

### Set up GPT Client

my_client = chadtools.setup_gpt_client()

### Use GPT to identify cuisines

<div class="alert alert-warning">
    <strong>WARNING:</strong> This step calls the OpenAI GPT API, which incurs costs. Ensure sufficient tokens are available before running this step.
</div>

df_filtered["cuisine"] = df_filtered["recipe_comment"].progress_apply(chadtools.get_cuisine, client=my_client)

### Filter out ingredients portion of the comments

We will use `regex` to extract the ingredients portion of the comment, and disregard the cooking instructions for simplicity. We will then use GPT-3.5 to identify the ingredients.

df_filtered['ingredient_comment'] = df_filtered['recipe_comment'].progress_apply(chadtools.extract_ingredients_text)

df_filtered['ingredient_comment'].head()

### Use GPT to identify ingredients

<div class="alert alert-warning">
    <strong>WARNING:</strong> This step calls the OpenAI GPT API, which incurs costs. Ensure sufficient tokens are available before running this step. Additionally, calling the API for a large number of recipes takes around 40 minutes.
</div>

df_filtered['gpt_ingredients'] = df_filtered['ingredient_comment'].progress_apply(chadtools.get_ingredient_list, client=my_client)
df_filtered['gpt_ingredients'].head()

### Reorder, filter, save as JSON

Finally, filter and reorder the columns, and save the dataframe as a JSON file for use in the next step.

ordered_cols = ["id", "title", "gpt_ingredients", "cuisine", "ingredient_comment", "score", "upvote_ratio", "link_flair_text", "author", "created_utc", "url", "permalink"]
df_filtered = df_filtered.loc[:, ordered_cols]
df_filtered.to_json('../data/cleaned_posts_with_cuisine_and_ingredient_list.json', orient='records', indent=4)
