==> notebooks/04 Analysis of Reddit Data.ipynb <==
# ✅Step 4. Analysis of Reddit Data

## 0. 🎯Import libraries

import pandas as pd
from plotnine import ggplot, aes, geom_histogram, geom_bar, facet_wrap, theme_minimal, ggtitle, themes, element_text, geom_bar, coord_flip, scale_x_discrete, labs
from plotnine.data import mtcars
from pprint import pprint

## 1. 🎯Load DataFrame

# Load the DataFrame from the file
filtered_df_posts = pd.read_json('../data/cleaned_posts_with_cuisine_and_ingredient_list.json')
filtered_df_posts.head()

flair_names = ['Recipe', 'Dessert', 'Pasta', 'Poultry', 'Fruit\\Vegetarian', 'Drink', 'Beef', 'Pork', 'Seafood']

## 2. 📊Plot data

We first plot a histogram of the upvote ratio of all the posts.

plot = (
    ggplot(
    filtered_df_posts, aes(x='upvote_ratio')) + 
    geom_histogram(binwidth=0.01, fill='#5c3da4') + 
    themes.theme_matplotlib() + 
    ggtitle("Upvote ratio for all posts") + 
    themes.theme(plot_title = element_text(weight='bold', color="black")) +
    themes.theme(aspect_ratio=9/16)
)
plot.save(f"../plots/plot_all_upvote_ratio.jpg", format="jpg", dpi=600)

# Save another copy to the ../docs directory
plot.save("../docs/plot_all_upvote_ratio.jpg", format="jpg", dpi=600)

We now separate the data by flair type to understand how well-received each type of recipe is.

for flair in flair_names:
    plot = (
        ggplot(filtered_df_posts[filtered_df_posts["link_flair_text"] == flair], aes(x='upvote_ratio')) + 
        geom_histogram(binwidth=0.01, fill='#5c3da4') +
        themes.theme_matplotlib() +
        ggtitle(f"Frequency of {flair} posts by upvote ratio") +
        themes.theme(plot_title = element_text(weight='bold', color="black")) +
        themes.theme(aspect_ratio=9/16)
    )
    plot.save(f"../plots/plot_{flair}_upvote_ratio.jpg".replace("\\", ""), format="jpg", dpi=600)

Another way to investigate the reception of different recipe types is to count the number of each type of recipe found in the most popular recipes. We will use the top 10% of posts by upvote ratio.

# Sort by upvote ratio in descending order
df_sorted_by_upvote_ratio = filtered_df_posts.sort_values(by='upvote_ratio', ascending=False)

# Calculate the number of rows for the top 10%
top_10_percent_by_upvote_ratio = int(0.1 * len(df_sorted_by_upvote_ratio))

# Take the top 10% of the DataFrame
top_10_df = df_sorted_by_upvote_ratio.head(top_10_percent_by_upvote_ratio)

ordered_flair_list = top_10_df['link_flair_text'].value_counts().index.tolist()
# Plot a bar graph showing the number of posts from different flairs
plot = (ggplot(top_10_df, aes(x="link_flair_text")) +
       geom_bar(fill='#5c3da4') +
       coord_flip() +
       ggtitle("Top 10% of posts by upvote ratio") +
       scale_x_discrete(limits=ordered_flair_list[::-1]) +
       themes.theme(plot_title = element_text(weight='bold', color="black")) +
       themes.theme(aspect_ratio=9/16) +
       labs(x="Flair", y="Number of posts")
)

plot.save(f"../plots/plot_top_10_percent_upvote_ratio.jpg", format="jpg", dpi=600)

# Save another copy to the ../docs directory
plot.save("../docs/plot_top_10_percent_upvote_ratio.jpg", format="jpg", dpi=600)

What about for absolute popularity? We repeat the same analysis but use the number of upvotes instead of the upvote ratio.

df_sorted_by_score = filtered_df_posts.sort_values(by='score', ascending=False)

# Calculate the number of rows for the top 10%
top_10_percent_by_score = int(0.1 * len(df_sorted_by_score))

# Take the top 10% of the DataFrame
top_10_df = df_sorted_by_score.head(top_10_percent_by_score)

ordered_flair_list = top_10_df['link_flair_text'].value_counts().index.tolist()
# Plot a bar graph showing the number of posts from different flairs
plot = (ggplot(top_10_df, aes(x="link_flair_text")) +
       geom_bar(fill='#5c3da4') +
       coord_flip() +
       ggtitle("Top 10% of posts by absolute score") +
       scale_x_discrete(limits=ordered_flair_list[::-1]) +
       themes.theme(plot_title = element_text(weight='bold', color="black")) +
       themes.theme(aspect_ratio=9/16) +
       labs(x="Flair", y="Number of posts")
)

plot.save(f"../plots/plot_top_10_percent_score.jpg", format="jpg", dpi=600)

Further elaboration on the plots can be found on the website.
