==> notebooks/05 Linking Reddit and BBC Good Food.ipynb <==
# ✅Step 5. Linking Reddit and BBC Good Food

# 🎯0. Import libraries

import requests               
import pandas as pd
from scrapy import Selector  
from tqdm import tqdm
from datetime import datetime
import sys
from plotnine import *
import altair as alt


from sentence_transformers import SentenceTransformer
from sklearn.metrics.pairwise import cosine_similarity


# Import our own modules
sys.path.append("../scripts/")
import chadtools

# 🎯1. Prepare GET requests for BBC GoodFood

bbclink = 'https://www.bbcgoodfood.com/search?q='
response = requests.get(bbclink)
sel = Selector(text=response.text)
links = []

recipecards = sel.css('main div.search-results div.card__section.card__content a ::attr(href)').getall()
links.extend(recipecards)  
prefix = "https://www.bbcgoodfood.com/recipes/"
final_links = [prefix + item for item in links] 

base_url = 'https://www.bbcgoodfood.com/search?q=&limit=1000&page='

links = []
for i in tqdm(range(1, 11)):
    response = requests.get(base_url + str(i))
    sel = Selector(text = response.text)
    recipecards = sel.css('main div.search-results div.card__section.card__content a ::attr(href)').getall()
    links.extend(recipecards)  

prefix = "https://www.bbcgoodfood.com/recipes/"
final_links = [prefix + item for item in links] 

len(final_links)

## 🎯2. Scrape BBC GoodFood Recipes

We will scrape all of BBC GoodFood's recipes and create a DataFrame containing the nutritional data of each of them, i.e. how much salt, fat, protein, etc. each recipe contains.

<div class="alert alert-warning">
    <strong>WARNING:</strong> This code block takes over 3 hours to run. For testing, use the pre-saved file <code>bbc_data.csv</code> in the next code block.
</div>

session = requests.Session()

def get_nutrition(url):
    response2 = session.get(url)
    sel2 = Selector(text=response2.text)
    table = sel2.css('table.key-value-blocks.hidden-print.mt-xxs')
    bbc = {}
    bbc['bbcgf_title'] = sel2.css('h1.heading-1 ::text').get()
    bbc['calories'] = table.css('td.key-value-blocks__value ::text').get()
    bbc['salt'] = table.xpath('.//*[contains(text(), "salt")]/..').css('td.key-value-blocks__value ::text').get()
    bbc['fat'] = table.xpath('.//*[contains(text(), "fat")]/..').css('td.key-value-blocks__value ::text').get()
    bbc['sugars'] = table.xpath('.//*[contains(text(), "sugars")]/..').css('td.key-value-blocks__value ::text').get()
    bbc['saturates'] = table.xpath('.//*[contains(text(), "saturates")]/..').css('td.key-value-blocks__value ::text').get()
    bbc['carbs'] = table.xpath('.//*[contains(text(), "carbs")]/..').css('td.key-value-blocks__value ::text').get()
    bbc['protein'] = table.xpath('.//*[contains(text(), "protein")]/..').css('td.key-value-blocks__value ::text').get()
    bbc['fibre'] = table.xpath('.//*[contains(text(), "fibre")]/..').css('td.key-value-blocks__value ::text').get()
    bbc['bbcgf_ratings_raw'] = sel2.css("div.rating__values span.sr-only ::text").get()
    return bbc

bbc = [get_nutrition(url) for url in tqdm(final_links)]

df_bbc = pd.DataFrame(bbc)
df_bbc['bbcgf_ratings'] = df_bbc['bbcgf_ratings_raw'].str.extract(r'(\d+\.\d+)').astype(float) / 5.0 
df_bbc.head(10)

### Save the data as a CSV file 

df_bbc.to_csv('../data/bbc_data.csv', index=False)

Save the data as HTML for website

df_bbc.head().to_html('../docs/bbc_data.html')

### Read cleaned_posts_with_cuisine_and_ingredient_list.json and bbc_data.csv

df_filtered = pd.read_json('../data/cleaned_posts_with_cuisine_and_ingredient_list.json', orient='records')
df_filtered.tail(3)

df_bbc = pd.read_csv('../data/bbc_data.csv')
df_bbc.head()

df_bbc.info()

Save a sample of the data as HTML for display on webpage

df_bbc.head().to_html('../docs/bbc_data.html', render_links=True, index=False)

## 🎯3. Merging the Reddit and BBC GoodFood dataframes 

How do we link the recipes from the two sites together? We use the following steps:
1. Calculate cosine similarity between the recipes from the two sites
2. If the cosine similarity is above our threshold (0.75), we consider the recipes to be the same.
    1. If there are multiple recipes that are above our threshold, we select the one with the highest similarity.
3. We then merge the two dataframes on the recipe title 

df_reddit = df_filtered.copy()

model = SentenceTransformer('sentence-transformers/all-mpnet-base-v2')
similarity_threshold = 0.75
merge_columns = ['bbcgf_title', 'calories', 'salt', 'fat', 'sugars', 'saturates', 'carbs', 'protein', 'fibre', 'bbcgf_ratings']

# Encode titles using Sentence Transformer
embeddings_reddit = model.encode(df_reddit['title'].tolist())
embeddings_bbc = model.encode(df_bbc['bbcgf_title'].tolist())

# Calculate cosine similarity matrix
cosine_sim_matrix = cosine_similarity(embeddings_reddit, embeddings_bbc)

# Find the indices and values of the maximum similarity in each row
max_similarity_indices = cosine_sim_matrix.argmax(axis=1)
max_similarity_values = cosine_sim_matrix.max(axis=1)

# Mask indices where the similarity is below the threshold
mask = max_similarity_values > similarity_threshold

# Merge columns based on the max similarity
df_reddit.loc[mask, merge_columns] = df_bbc.iloc[max_similarity_indices[mask]][merge_columns].values

df_reddit.tail()

We drop rows with no nutritional data to get a clean dataframe for further analysis.

df_for_analysis = df_reddit.dropna(subset=merge_columns, how='any').copy()
df_for_analysis.info()

df_for_analysis.iloc[:, 13:] = df_for_analysis.iloc[:, 13:].astype("float16")
df_for_analysis["score"] = df_for_analysis["score"].astype("int16")
df_for_analysis["upvote_ratio"] = df_for_analysis["upvote_ratio"].astype("float16")

df_for_analysis.info()

df_for_analysis.head()

df_for_analysis.to_json('../data/merged_data_for_analysis.json', orient='records', indent=4)

Saving as html for website

df_for_analysis.head().to_html('../docs/merged_data_for_analysis.html', render_links=True, index=False)

## Analysis of merged Reddit and BBC Good Food data

### Plotting histograms of BBC Good Food ratings against each flair 
Having obtained the user ratings from BBC Good Food, we can compare if there are any differences between the category of foods that are popular amongst Reddit vs BBC Good Food users. 

df_for_analysis = pd.read_json('../data/merged_data_for_analysis.json', orient='records')

df_desserts = df_for_analysis[df_for_analysis['link_flair_text']=='Dessert']
df_recipe = df_for_analysis[df_for_analysis['link_flair_text']=='Recipe']
df_pasta = df_for_analysis[df_for_analysis['link_flair_text']=='Pasta']
df_poultry = df_for_analysis[df_for_analysis['link_flair_text']=='Poultry']
df_drink = df_for_analysis[df_for_analysis['link_flair_text']=='Drink']
df_beef = df_for_analysis[df_for_analysis['link_flair_text']=='Beef']
df_pork = df_for_analysis[df_for_analysis['link_flair_text']=='Pork']
df_seafood = df_for_analysis[df_for_analysis['link_flair_text']=='Seafood']
df_fruitveg = df_for_analysis[df_for_analysis['link_flair_text']=='Fruit\Vegetarian']

# plot a histogram of the bbc ratings for each flair category
flair_names = ['Dessert', 'Recipe', 'Pasta', 'Poultry', 'Drink', 'Beef', 'Pork', 'Seafood', 'Fruit\Vegetarian']

for flair in flair_names:
    plot = (
        ggplot(df_for_analysis[df_for_analysis["link_flair_text"] == flair], aes(x='bbcgf_ratings')) + 
        geom_histogram(binwidth=0.01, fill='#5c3da4') +
        theme_matplotlib() + 
        ggtitle(f"Frequency of {flair} posts by BBC Good Food ratings ratio") +
        theme(plot_title = element_text(weight='bold', color="black")) +
        theme(aspect_ratio=9/16)
    )

    plot.save(f"../plots/{flair}_hist.png".replace("\\", ""), dpi=300) 

### Plot histogram of top 10 percent of BBC GoodFood recipes by website ratings

# Sort by 'upvote ratio' in descending order
df_sorted = df_for_analysis.sort_values(by='bbcgf_ratings', ascending=False)

# Calculate the number of rows for the top 10%
top_10_percent = int(0.1 * len(df_sorted))

# Take the top 10% of the DataFrame
top_10_df = df_sorted.head(top_10_percent)

ordered_flair_list = top_10_df['link_flair_text'].value_counts().index.tolist()
# Plot a bar graph showing the number of posts from different flairs
plot = (
       ggplot(top_10_df, aes(x="link_flair_text")) + 
       geom_bar(fill='#5c3da4') +
       coord_flip() +
       ggtitle("Top 10% of posts by BBC Good Food ratings") +
       scale_x_discrete(limits=top_10_df["link_flair_text"].value_counts().index.tolist()[::-1]) +
       theme(plot_title = element_text(weight='bold', color="black")) +
       theme(aspect_ratio=9/16)
       )

plot.save(f"../plots/plot_top_10_percent_bbcgf.jpg", format="jpg", dpi=600)

### Plotting an interactive scatter plot of calories against upvote ratio

For easier visualisation, we combine the 'Poultry', 'Beef', 'Pork', 'Seafood' flairs into one flair titled 'Meats'.

df_for_altair = df_for_analysis.copy()

df_for_altair['link_flair_text'].replace(['Poultry', 'Beef', 'Pork', 'Seafood'], 'Meats', inplace=True)

x_scale = alt.Scale(domain=(0.5, 1))
colour_scale = alt.Scale(domain=['Dessert', 'Recipe', 'Pasta', 'Meats', 'Drink', 'Fruit\\Vegetarian'],
                         range=['#41afaa', '#466eb4', '#00a0e1', '#e6a532', '#d7642c', '#af4b91'])

selection = alt.selection_point(fields=['link_flair_text'], bind='legend')

scatter = alt.Chart(df_for_altair, width=600, height=600).mark_circle().encode(
    x=alt.X('upvote_ratio:Q', scale=x_scale, title='Upvote Ratio'),
    y=alt.Y('calories:Q', title='Calories'), 
    color=alt.Color('link_flair_text:N', scale=colour_scale, legend=alt.Legend(title='Flair')),
    tooltip=['title:N', 'upvote_ratio:Q', 'link_flair_text:N', 'calories:Q', 'salt:Q', 'fat:Q', 'sugars:Q', 'saturates:Q', 'carbs:Q', 'protein:Q', 'fibre:Q', 'permalink:N'],
    opacity=alt.condition(selection, alt.value(1), alt.value(0.15)),
    href='permalink:N'
).add_params(
    selection
).properties(title='No clear correlation between the healthiness of a recipe and how well it is received on Reddit',
).interactive()

scatter.save('../docs/interactive_plot.html', scale_factor=2.0)

We then plot an interactive scatter plot of reddit upvote ratios against BBC GoodFood ratings. Unsurprisingly, higher rated recipes on BBC GoodFood tend to have higher upvote ratios on Reddit.

x_scale = alt.Scale(domain=(0.5, 1))
y_scale = alt.Scale(domain=(0.5, 1))

scatter = alt.Chart(df_for_altair, width=600, height=600).mark_circle().encode(
    x=alt.X('upvote_ratio:Q', scale=x_scale, title='Upvote Ratio'),
    y=alt.Y('bbcgf_ratings:Q', scale=y_scale, title='BBC Good Food Rating'),
    tooltip=['title:N', 'upvote_ratio:Q', 'bbcgf_ratings:Q', 'permalink:N'],
    href='permalink:N'
).properties(title='Higher Reddit upvote ratios are generally associated with higher BBC Good Food ratings'
).interactive()

scatter.save('../docs/upvote_ratio_vs_bbcgf_rating.html', scale_factor=2.0) 

Further analysis of specific ingredients and cuisines will be done in the next notebook.
