==> notebooks/06 Cuisine Analysis.ipynb <==
# ✅Step 6: Analysis of Ingredients and Cuisines

## 🎯0. Import libraries and load data

import pandas as pd
import numpy as np
import sys
from plotnine import *
import altair as alt

# Import our own modules
sys.path.append("../scripts/")
import chadtools

df = pd.read_json("../data/merged_data_for_analysis.json", orient="records")
df.head(3)

## 🎯1. Analysis of Ingredients

We first do some ingredient analysis for all the recipes. In particular, we are interested in finding how much of the *top 10 most common ingredients* are used in each cuisine.

all_ingredient_list = []
for i in df["gpt_ingredients"].tolist():
    all_ingredient_list.extend(i)

all_ingredient_series = pd.Series(all_ingredient_list)
top_10_ingredients = all_ingredient_series.value_counts().head(10)

top_10_ingredients

ingredient_frequency_df = pd.DataFrame(top_10_ingredients).reset_index()
ingredient_frequency_df.columns = ["ingredient", "frequency"]
ingredient_frequency_df = ingredient_frequency_df.sort_values("frequency", ascending=False)

plot = (ggplot(ingredient_frequency_df, aes(x="ingredient", y="frequency")) +
       geom_bar(stat="identity", fill='#5c3da4') +
       coord_flip() +
       scale_x_discrete(limits=ingredient_frequency_df["ingredient"][::-1]) +
       theme(plot_title = element_text(weight='bold', color="black", size=10, ha="right")) +
       theme(aspect_ratio=9/16) +
       labs(title="Top 10 ingredients by frequency in both BBC GoodFood and Reddit recipes", 
            x="Ingredient", y="Number of recipes")
)
plot.save("../plots/top_10_ingredients_frequency.jpg", format="jpg", dpi=600)

Hence, the top 10 ingredients found in recipes appearing on both BBC GoodFood and Reddit are:
1. Salt
2. Pepper
3. Garlic
4. Butter
5. Sugar
6. Olive oil
7. Water
8. Black Pepper
9. Eggs
10. Yellow Onion

Save the dataframe as a HTML for webpage

ingredient_frequency_df.to_html("../docs/ingredient_frequency.html", index=False)

## 🎯2. Analysis of Cuisines

We are also interested in the breakdown of cuisines in the recipes.

# get the top 10 cuisines by frequency 
cuisine_counts = df["cuisine"].value_counts().reset_index()
cuisine_counts.columns = ["cuisine", "count"]
cuisine_counts = cuisine_counts.sort_values("count", ascending=False)
cuisine_counts.head(10)

# if the recipe is in the top 10, keep it, otherwise set it to "other"
df["cuisine"] = df["cuisine"].apply(lambda x: x if x in cuisine_counts["cuisine"].values[:10] else "other")

df.head()

## 🎯3. Putting both together

### 3.1 Frequency of top 10 ingredients in different cuisines

What ingredients feature prominently in each cuisine? We will put the top 10 ingredients and cuisines together to find out. We group the recipes by cuisines and find the mean quantity of each nutrient for the top 10 cuisines.

# for each ingredient, get the frequency of each cuisine
cuisine_ingredient_df = pd.DataFrame()
for ingredient in ingredient_frequency_df["ingredient"]:
    ingredient_df = df[df["gpt_ingredients"].apply(lambda x: ingredient in x)]
    cuisine_counts = ingredient_df["cuisine"].value_counts().reset_index()
    cuisine_counts.columns = ["cuisine", "count"]
    cuisine_counts["ingredient"] = ingredient
    cuisine_ingredient_df = pd.concat([cuisine_ingredient_df, cuisine_counts])
    cuisine_ingredient_df = cuisine_ingredient_df.reset_index(drop=True)

cuisine_ingredient_df

custom_colors = {
    'american': '#084C61',  
    'british': '#EB4B98',  
    'chinese': '#FFC300',
    'french': '#43B3AE',   
    'greek': '#614051',  
    'indian': '#4F7942',   
    'italian': '#0F52BA',
    'japanese': '#F28500',  
    'mexican': '#D21F3C', 
    'thai': '#B87333', 
    'other': '#708090' 
}

plot = (ggplot(cuisine_ingredient_df, aes(x="ingredient", y="count", fill="cuisine")) +
       geom_bar(stat="identity") +
       coord_flip() +
       scale_x_discrete(limits=ingredient_frequency_df["ingredient"][::-1]) +
       scale_fill_manual(values=custom_colors) +
       theme_minimal() +
       theme(axis_text_x=element_text(angle=90)) +
       labs(title="Top 10 ingredients by frequency in each cuisine", x="Ingredient", y="Number of recipes") +
       theme(aspect_ratio=7/11)
)  
plot.save("../docs/top_10_ingredients_cuisine.jpg", format="jpg", dpi=600)

# cross tabulate the data to show the frequency of each ingredient in each cuisine
final_cuisine_ingredient_df = cuisine_ingredient_df.pivot(index="cuisine", columns="ingredient", values="count")
final_cuisine_ingredient_df = final_cuisine_ingredient_df.fillna(0)
final_cuisine_ingredient_df = final_cuisine_ingredient_df.astype(int)

# Reset column index to remove "ingredient" label
final_cuisine_ingredient_df.columns.name = None
final_cuisine_ingredient_df = final_cuisine_ingredient_df.reset_index()

final_cuisine_ingredient_df

### 3.2 Mean nutrition data for top 10 cuisines 

nutrients = ["calories", "carbs", "fat", "protein", "saturates", "sugars", "fibre", "salt"]
top_10_cuisines = df["cuisine"].value_counts().head(10)
healthy_nutrient_amounts = [750, 40, 20, 19, 10, 10, 10, 2]
healthy_nutrient_dict = {
    "cuisine": " healthy threshold",
    "calories": 750,
    "carbs": 40,
    "fat": 20,
    # "protein": 19,
    "saturates": 10,
    "sugars": 10,
    # "fibre": 10,
    "salt": 2
}

cuisines_nutrients_df = df[["cuisine"] + list(healthy_nutrient_dict.keys())[1:]].groupby(["cuisine"]).mean().filter(top_10_cuisines.index, axis=0).reset_index()
cuisines_nutrients_df

We also set a healthy threshold for each nutrient based on BBC GoodFood and NHS guidelines. This will motivate our health comparisons.

new_row = pd.DataFrame(healthy_nutrient_dict, index=[0])
new_row

cuisines_nutrients_df = pd.concat([cuisines_nutrients_df, new_row]).reset_index(drop=True)
cuisines_nutrients_df

To plot a radar chart, will use the `ggradar` library in R. To show the relative amount of each ingredient among the cuisines, we need to normalise the quantities with a min-max scaling method.

cuisines_nutrients_df_normalised = cuisines_nutrients_df.iloc[:, :2].copy()

for nutrient in list(healthy_nutrient_dict.keys())[1:]:
    series = cuisines_nutrients_df[nutrient]
    normalized_series = (series - series.min()) / (series.max() - series.min()) # apply min-max scaling
    cuisines_nutrients_df_normalised[nutrient] = normalized_series # add normalised column

# Capitalise cuisine names
cuisines_nutrients_df_normalised["cuisine"] = cuisines_nutrients_df_normalised["cuisine"].apply(lambda x: x.capitalize())

cuisines_nutrients_df_normalised

### Save to CSV to use in R

cuisines_nutrients_df_normalised.to_csv("../data/top10cuisines_nutrients_normalised.csv", index=False)

The code for visualisation can be found in [radar_plots.r](../scripts/radar_plots.r)
